Add dark mode inheritance
