+++ ios/MapNews/Client/NewsClient.swift
