+++ ios/MapNews/Model/MapViewModel.swift
