+++ ios/MapNews/Util/StringExtension.swift
